Sheep + Packages + Collecting health and coins

+++ SheepToss/Assets/Scripts/ILivable.cs
@@ -4,5 +4,5 @@ using System;
 
 interface ILivable
 {
-    int HP { get; set; }
+    float HP { get; set; }
 }
